+++ packages/edit-site/src/store/private-selectors.js
@@ -1,4 +1,9 @@
 /**
+ * Internal dependencies
+ */
+import { hasPageContentFocus } from './selectors';
+
+/**
  * Returns the current canvas mode.
  *
  * @param {Object} state Global application state.
@@ -18,4 +23,21 @@ export function getCanvasMode( state ) {
  */
 export function getEditorCanvasContainerView( state ) {
 	return state.editorCanvasContainerView;
+}
+
+/**
+ * Returns the type of the current page content focus, or null if there is no
+ * page content focus.
+ *
+ * Possible values are:
+ *
+ * - `'disableTemplate'`: Disable the blocks belonging to the page's template.
+ * - `'hideTemplate'`: Hide the blocks belonging to the page's template.
+ *
+ * @param {Object} state Global application state.
+ *
+ * @return {'disableTemplate'|'hideTemplate'|null} Type of the current page content focus.
+ */
+export function getPageContentFocusType( state ) {
+	return hasPageContentFocus( state ) ? state.pageContentFocusType : null;
 }
